Harden legacy GC counter width and sleep parking for the Windows target
Two review findings on the Windows (LLP64 / no-signal-stop) target:

- cn1LegacyBytesSinceGc was _Atomic long, which is 32-bit under LLP64;
  the counter accumulates raw bytes between cycle resets, so it now uses
  long long (with the fetch_add operand and prevLegacyBytes widened to
  match) so the edge-trigger comparison can never see a wrapped value.

- Thread.sleep published threadActive=FALSE without a cooperative park
  capture (a pre-existing pattern this PR's rewrite kept). Under
  CN1_CONSERVATIVE_GC_ROOTS the collector scans a parked thread's native
  stack either via the capture or a signal stop, and Windows has no
  signal-stop fallback -- an uncaptured sleeper's native-stack roots
  went unscanned for the cycle. sleep now runs CN1_GC_PARK_CAPTURE
  before parking, per the CN1_YIELD_THREAD contract, and drops the
  capture on resume exactly like CN1_RESUME_THREAD.

Revalidated locally: LargeArrayLoad 6 GC cycles (x2), RESULT=1099 host
parity, SLEEP_EDGE_OK, TIMER_LATENCY_OK (max_early_ms=0,
min_sleep1500_ms=1500), MapTorture byte-identical to the host JVM,
GcStress x4 + MtStress x3 in cooperative and CN1_GC_SIGNAL_STOP=1 modes.

+++ vm/ByteCodeTranslator/src/cn1_globals.m
@@ -1818,7 +1818,9 @@ long long cn1_instr_allocCount = 0;
 #ifndef CN1_LEGACY_GC_TRIGGER_BYTES
 #define CN1_LEGACY_GC_TRIGGER_BYTES (24*1024*1024)
 #endif
-static _Atomic long cn1LegacyBytesSinceGc = 0;
+// long long, NOT long: on the Windows LLP64 target long is 32-bit and this
+// counter accumulates raw allocation bytes between cycle resets.
+static _Atomic long long cn1LegacyBytesSinceGc = 0;
 #endif
 
 JAVA_BOOLEAN java_lang_System_isHighFrequencyGC___R_boolean(CODENAME_ONE_THREAD_STATE) {
@@ -3959,10 +3961,11 @@ JAVA_OBJECT codenameOneGcMalloc(CODENAME_ONE_THREAD_STATE, int size, struct claz
     // (which is why there is deliberately no !gcCurrentlyRunning suppression:
     // suppressing would LOSE the edge and delay collection up to the 30s
     // idle wait).
-    long prevLegacyBytes = atomic_fetch_add_explicit(&cn1LegacyBytesSinceGc, (long)size,
-                                                     memory_order_relaxed);
+    long long prevLegacyBytes = atomic_fetch_add_explicit(&cn1LegacyBytesSinceGc,
+                                                          (long long)size,
+                                                          memory_order_relaxed);
     if(prevLegacyBytes <= CN1_LEGACY_GC_TRIGGER_BYTES
-       && prevLegacyBytes + (long)size > CN1_LEGACY_GC_TRIGGER_BYTES
+       && prevLegacyBytes + (long long)size > CN1_LEGACY_GC_TRIGGER_BYTES
        && constantPoolObjects != 0
 #ifndef CN1_CONSERVATIVE_GC_ROOTS
        // Mirror cn1BibopMaybeGc's gate exactly: without conservative roots a
